Update to-do_chore_notifications.yaml
Update To-do Chore Notifications with the optional ability to run actions after a chore is marked complete. There is also a variable called "todo_item" which can be used in those actions, and it the name of the chore that was completed. This allows running different actions based on the name of the completed chore.